% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_summary.R
\name{col_summary}
\alias{col_summary}
\title{Column summaries}
\usage{
col_summary(df, fun = c("mean", "sd", "min", "max"))
}
\arguments{
\item{df}{A data.frame, array or matrix.}

\item{fun}{The summary to be applied. One of "mean", "sd", "min", "max".}
}
\value{
Named vector with per column summaries.
}
\description{
Column summaries
}
