% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot summaries as bar chart}
\usage{
plot_summary(vec)
}
\arguments{
\item{vec}{A named vector with summary statistics.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot summaries as bar chart
}
